backend/requirements.txt
Django==2.2.19
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
flake8==4.0.1
mccabe==0.6.1
pycodestyle==2.8.0
pyflakes==2.4.0
PyJWT==2.4.0
python-dotenv==0.20.0
pytz==2022.1
sqlparse==0.4.2
Pillow==9.1.1